role_base

+++ src/components/admin/brand/modal/ConfirmDeleteBrand.vue
@@ -13,9 +13,18 @@ const { confirmFlag, brandName, brandId } = storeToRefs(store)
 const { SET_CONFIRM_FLAG } = store
 
 const handleDeleteCategory = async () => {
+    let roleLocal = localStorage.getItem("role")
+    let role = roleLocal ? JSON.parse(roleLocal) : ''
+
+    if (role !== "admin") {
+        toast.error("You are not an ADMIN")
+        SET_CONFIRM_FLAG(false)
+        return
+    }
+
     let response = await api.deleteBrand(brandId.value)
 
-    if(!response){
+    if (!response) {
         toast.error("Brand is deleted failed")
         return
     }
